Added Keyboard Preferences panel and updated keyboard help.
Keyboard Preferences panel allows a user to toggle the hotkey
suppression, and offers a button to open the Universal Access
preferences if they need to "Enable access for assistive devices" there
also.

Also made hotkey suppression only take effect while a program is
running and the emulation is not paused.